+++ geolocator/ios/Classes/Utils/LocationMapper.m
@@ -13,16 +13,30 @@
         return nil;
     }
     
-    return @{
-        @"latitude": @(location.coordinate.latitude),
-        @"longitude": @(location.coordinate.longitude),
-        @"timestamp": @([LocationMapper currentTimeInMilliSeconds:location.timestamp]),
-        @"altitude": @(location.altitude),
-        @"accuracy": @(location.horizontalAccuracy),
-        @"speed": @(location.speed),
-        @"speed_accuracy": @0.0,
-        @"heading": @(location.course),
-    };
+    double timestamp = [LocationMapper currentTimeInMilliSeconds:location.timestamp];
+    double speedAccuracy = 0.0;
+    
+#if __IPHONE_OS_VERSION_MAX_ALLOWED >= 140000
+    if (@available(iOS 10.0, *)) {
+        speedAccuracy = location.speedAccuracy;
+    }
+#endif
+    
+    NSMutableDictionary *locationMap = [[NSMutableDictionary alloc]initWithCapacity:9];
+    [locationMap setObject:@(location.coordinate.latitude) forKey: @"latitude"];
+    [locationMap setObject:@(location.coordinate.longitude) forKey: @"longitude"];
+    [locationMap setObject:@(timestamp) forKey: @"timestamp"];
+    [locationMap setObject:@(location.altitude) forKey: @"altitude"];
+    [locationMap setObject:@(location.horizontalAccuracy) forKey: @"accuracy"];
+    [locationMap setObject:@(location.speed) forKey: @"speed"];
+    [locationMap setObject:@(speedAccuracy) forKey: @"speed_accuracy"];
+    [locationMap setObject:@(location.course) forKey: @"heading"];
+    
+    if(location.floor && location.floor.level) {
+        [locationMap setObject:@(location.floor.level) forKey:@"floor"];
+    }
+    
+    return locationMap;
 }
 
 + (double)currentTimeInMilliSeconds:(NSDate *)dateToConvert {