+++ geolocator_platform_interface/lib/src/geolocator_platform_interface.dart
@@ -75,7 +75,7 @@ abstract class GeolocatorPlatform extends PlatformInterface {
   /// When no position is available, null is returned.
   /// Throws a [PermissionDeniedException] when trying to request the device's
   /// location when the user denied access.
-  Future<Position> getLastKnownPosition({
+  Future<Position?> getLastKnownPosition({
     bool forceAndroidLocationManager = false,
   }) {
     throw UnimplementedError(